pyproject.toml
[tool.poetry]
name = "cf-template"
version = "0.1.0"
description = ""
authors = ["vetlenev <vetle.nevland@akerbp.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^1.26.2"
pandas = "^2.1.3"
python-dotenv = "^1.0.0"
cognite-sdk = "^7.5.1"
ipykernel = "^6.27.1"
pytest = "^7.4.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
